moved ResizeNode into its own file
