+++ src/lib/controllers/util.ts
@@ -9,7 +9,7 @@ import type {
   UserNodeType,
   UserEdgeType,
 } from '$lib/models/types';
-import { ResizeNode } from '$lib/models/store';
+import { ResizeNode } from '$lib/models/ResizeNode';
 import { Anchor } from '$lib/models/Anchor';
 import { Node } from '$lib/models/Node';
 import { Edge } from '$lib/models/Edge';
